Merge in OMNI/lsmag-two from OMNI-5538 to master

* commit '68e8d7db01162a14bdc73562ccdca5e391770b45':
  [OMNI-5538] For fixing the formatting of loyalty points earned while creating order from admin

+++ src/Omni/Controller/Ajax/UpdatePoints.php
@@ -114,7 +114,7 @@ class UpdatePoints extends Action
         }
         $post          = $this->getRequest()->getContent();
         $postData      = json_decode($post);
-        $loyaltyPoints = (int)$postData->loyaltyPoints;
+        $loyaltyPoints = (float)$postData->loyaltyPoints;
         $isPointValid  = $this->loyaltyHelper->isPointsAreValid($loyaltyPoints);
         if (!is_numeric($loyaltyPoints) || $loyaltyPoints < 0 || !$isPointValid) {
             $response = [
